refactor: Outbox/Inbox 선택적 활성화 지원
