+++ src/main/java/com/firstticket/common/messaging/scheduler/OutboxCleanupScheduler.java
@@ -1,0 +1,31 @@
+package com.firstticket.common.messaging.scheduler;
+
+import com.firstticket.common.messaging.inbox.QInbox;
+import com.firstticket.common.messaging.outbox.OutboxStatus;
+import com.firstticket.common.messaging.outbox.QOutbox;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDateTime;
+
+@Slf4j
+@RequiredArgsConstructor
+public class OutboxCleanupScheduler {
+
+    private final JPAQueryFactory queryFactory;
+
+    @Transactional
+    @Scheduled(cron = "0 0 3 * * *")
+    public void cleanupOutbox() {
+        long deleted = queryFactory
+            .delete(QOutbox.outbox)
+            .where(QOutbox.outbox.status.eq(OutboxStatus.PUBLISHED)
+                .and(QOutbox.outbox.publishedAt.before(LocalDateTime.now().minusWeeks(1L))))
+            .execute();
+        log.info("[Outbox] {}건 삭제 완료", deleted);
+    }
+
+}
